+++ internal/platforms/youtube/subscriptions/lease.go
@@ -1,4 +1,5 @@
-package store
+// file name: internal/platforms/youtube/store/lease.go
+package subscriptions
 
 import (
 	"errors"