+++ internal/platforms/youtube/subscriptions/lease_test.go
@@ -1,4 +1,4 @@
-package store
+package subscriptions
 
 import (
 	"path/filepath"